+++ src/main/kotlin/org/soneira/savings/api/dto/MovementDTO.kt
@@ -1,9 +1,8 @@
 package org.soneira.savings.api.dto
 
-import java.math.BigInteger
 
 data class MovementDTO(val id: String) {
     var description: String? = null
-    var subcategory: BigInteger? = null
+    var subcategory: Int? = null
     var comment: String? = null
 }
