+++ src/main/kotlin/org/soneira/savings/domain/model/vo/id/CategoryId.kt
@@ -1,5 +1,3 @@
 package org.soneira.savings.domain.model.vo.id
 
-import java.math.BigInteger
-
-class CategoryId(value: BigInteger) : BaseId<BigInteger>(value)
+class CategoryId(value: Int) : BaseId<Int>(value)
