+++ src/main/kotlin/org/soneira/savings/domain/repository/CategoryRepository.kt
@@ -1,7 +1,6 @@
 package org.soneira.savings.domain.repository
 
 import org.soneira.savings.domain.model.entity.Category
-import java.math.BigInteger
 
 
 interface CategoryRepository {
@@ -16,6 +15,6 @@ interface CategoryRepository {
      * @param id the category identifier
      * @return the category [Category]
      */
-    fun getById(id: BigInteger): Category
+    fun getById(id: Int): Category
 
 }
