+++ src/main/kotlin/org/soneira/savings/domain/repository/SubcategoryRepository.kt
@@ -1,7 +1,6 @@
 package org.soneira.savings.domain.repository
 
 import org.soneira.savings.domain.model.entity.Subcategory
-import java.math.BigInteger
 
 interface SubcategoryRepository {
     /**
@@ -15,6 +14,6 @@ interface SubcategoryRepository {
      * @param id the subcategory identifier
      * @return the subcategory [Subcategory]
      */
-    fun getById(id: BigInteger): Subcategory
+    fun getById(id: Int): Subcategory
 
 }
