+++ src/main/kotlin/org/soneira/savings/infrastructure/persistence/mongo/document/SubcategoryDocument.kt
@@ -3,6 +3,7 @@ package org.soneira.savings.infrastructure.persistence.mongo.document
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
 import org.springframework.data.mongodb.core.mapping.Field
+import org.springframework.data.mongodb.core.mapping.FieldType.INT32
 import org.springframework.data.mongodb.core.mapping.FieldType.INT64
 import java.math.BigInteger
 
@@ -14,6 +15,6 @@ data class SubcategoryDocument(
     val description: String,
     @Field("description_es")
     val descriptionEs: String,
-    @Field(targetType = INT64)
-    val category: BigInteger
+    @Field(targetType = INT32)
+    val category: Int
 )
